rangedoppler: hw_only: part2_2dMatchedFilterFFT Updated example with HDL simulink model Added a simulink HDL model that does matched filter with a transpose followed by an FFT. Some other notes: * Radar cube size is 256x64 * Fixed point values were expanded to fi(1,20,18) * Radar targets were repositioned such that they fit in the 256x64 view

+++ rangedoppler/hw_only/part2_2dMatchedFilterFFT/createRadarDataCubeMod.m
@@ -51,7 +51,7 @@ for nn=1:NumCPI
         [~,ang] = rangeangle(TgtPos);
         temp = rxCollector(temp,ang);
         temp = rx(temp);
-        dataCube(:,ii,nn) = temp(RngGateSamples+1:end,:);
+        dataCube(:,ii,nn) = temp(uint32(RngGateSamples+1):end,:);
     end
     T = T+CPIDelta;
 end
